feat: adiciona validacaoes com tratamentos de excecao e cors configs

+++ src/main/java/br/com/barbosa/transaction/services/TransactionService.java
@@ -2,6 +2,8 @@ package br.com.barbosa.transaction.services;
 
 import br.com.barbosa.transaction.entities.Category;
 import br.com.barbosa.transaction.entities.Transaction;
+import br.com.barbosa.transaction.exceptions.TransactionNotFoundException;
+import br.com.barbosa.transaction.exceptions.TransactionValidationException;
 import br.com.barbosa.transaction.repositories.CategoryRepository;
 import br.com.barbosa.transaction.repositories.TransactionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,35 +26,37 @@ public class TransactionService {
 
     public Transaction getTransactionById(String id) {
         if (id == null) {
-            throw new IllegalArgumentException("The given id must not be null");
+            throw new TransactionValidationException("O ID fornecido não deve ser nulo");
         }
         return transactionRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Transaction not found"));
+                .orElseThrow(() -> new TransactionNotFoundException("Transação não encontrada para o ID: " + id));
     }
 
     public Transaction saveTransaction(Transaction transaction) {
         if (transaction.getCategoryId() == null || transaction.getCategoryId().isEmpty()) {
-            throw new IllegalArgumentException("The given category id must not be null or empty");
+            throw new TransactionValidationException("O ID da categoria fornecido não deve ser nulo ou vazio");
         }
 
         Category category = categoryRepository.findById(transaction.getCategoryId())
-                .orElseThrow(() -> new RuntimeException("Category not found"));
+                .orElseThrow(() -> new TransactionNotFoundException("Categoria não encontrada para o ID: " + transaction.getCategoryId()));
 
         transaction.setCategory(category);
-
         return transactionRepository.save(transaction);
     }
 
     public void deleteTransaction(String id) {
         if (id == null) {
-            throw new IllegalArgumentException("O id fornecido não deve ser nulo");
+            throw new TransactionValidationException("O ID fornecido não deve ser nulo");
+        }
+        if (!transactionRepository.existsById(id)) {
+            throw new TransactionNotFoundException("Transação não encontrada para o ID: " + id);
         }
         transactionRepository.deleteById(id);
     }
 
     public Transaction updateTransaction(String id, Transaction updatedTransaction) {
         Transaction existingTransaction = transactionRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Transação não encontrada"));
+                .orElseThrow(() -> new TransactionNotFoundException("Transação não encontrada para o ID: " + id));
 
         existingTransaction.setDate(updatedTransaction.getDate());
         existingTransaction.setTitle(updatedTransaction.getTitle());
@@ -61,7 +65,7 @@ public class TransactionService {
 
         if (updatedTransaction.getCategoryId() != null) {
             Category category = categoryRepository.findById(updatedTransaction.getCategoryId())
-                    .orElseThrow(() -> new RuntimeException("Categoria não encontrada"));
+                    .orElseThrow(() -> new TransactionNotFoundException("Categoria não encontrada para o ID: " + updatedTransaction.getCategoryId()));
             existingTransaction.setCategory(category);
         }
